Samsung SMSC: Convert PDU SMSC into phone number format
The patch [1] parses the SMSC provided by Samsung's RIL from, for example:

"0791451115470055",0 to 0791451115470055

This a problem because Android seems to be fine with the "SMSC",format_specifier,
but it wants the SMSC as a phone number instead of PDU.

This patch makes it further decode the SMSC so that:

0791451115470055 becomes +541151740055

And that finally fixes SMS on my carrier (tested on Personal AR). Whether
this works with other carriers needs to be tested.

Checking stock ROM screenshots and other devices confirms that this is the
correct SMSC format.

[1] https://github.com/TrebleDroid/platform_frameworks_opt_telephony/commit/9de26ca6b0fcbbb8bc27ca75a57697a7724f797d

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>